top-level keys module
as keys API will likely be shared between credit and debit parts, this
commit creates a top-level keys module.
As a first step on the migration of crate::credit::keys APIs to
crate::keys module, we move struct
