+++ src/credit/keys.rs
@@ -9,6 +9,7 @@ use cdk::nuts::nut02 as cdk02;
 use thiserror::Error;
 // ----- local modules
 // ----- local imports
+use crate::keys::KeysetID;
 
 pub type Result<T> = std::result::Result<T, Error>;
 #[derive(Debug, Error)]
@@ -26,64 +27,7 @@ pub enum Error {
     #[error("repository error {0}")]
     Repository(#[from] Box<dyn std::error::Error>),
 }
-/// rework of cdk02::Id as they do not export internal fields
-#[derive(Debug, Clone, Copy, Hash, Eq, PartialEq)]
-pub struct KeysetID {
-    pub version: cdk02::KeySetVersion,
-    pub id: [u8; Self::BYTELEN],
-}
 
-impl KeysetID {
-    pub const BYTELEN: usize = 7;
-
-    pub fn new(bill: &str, endorser: &str) -> Self {
-        let input = format!("{}{}", bill, endorser);
-        let digest = Sha256::hash(input.as_bytes());
-        Self {
-            version: cdk02::KeySetVersion::Version00,
-            id: digest.as_byte_array()[0..Self::BYTELEN]
-                .try_into()
-                .expect("cdk::KeysetID BYTELEN == 7"),
-        }
-    }
-
-    pub fn as_bytes(&self) -> [u8; Self::BYTELEN + 1] {
-        let mut bytes = [0u8; Self::BYTELEN + 1];
-        bytes[0] = self.version as u8;
-        bytes[1..].copy_from_slice(&self.id);
-        bytes
-    }
-}
-
-impl std::cmp::PartialEq<cdk02::Id> for KeysetID {
-    fn eq(&self, other: &cdk02::Id) -> bool {
-        other.as_bytes() == self.as_bytes()
-    }
-}
-
-impl std::convert::From<cdk02::Id> for KeysetID {
-    fn from(id: cdk02::Id) -> Self {
-        let bb = id.to_bytes();
-        assert_eq!(bb.len(), Self::BYTELEN + 1);
-        assert_eq!(bb[0], cdk02::KeySetVersion::Version00.to_byte());
-        Self {
-            version: cdk02::KeySetVersion::Version00,
-            id: bb[1..].try_into().expect("cdk::KeysetID BYTELEN == 7"),
-        }
-    }
-}
-
-impl std::convert::From<KeysetID> for cdk02::Id {
-    fn from(id: KeysetID) -> Self {
-        Self::from_bytes(&id.as_bytes()).expect("cdk::KeysetID BYTELEN == 7")
-    }
-}
-
-impl std::fmt::Display for KeysetID {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}{}", self.version, hex::encode(self.id))
-    }
-}
 
 // ---------- Keys Repository for creation
 #[cfg_attr(test, mockall::automock)]
